+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,21 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+* Vendor new SDK and bump API version to 2022-11-02-preview.
+
+0.5.121
++++++++
 * Remove defender related code after GA, reuse the implementation in azure-cli/acs.
+* Remove check_raw_parameters in update code path, reuse the implementation in azure-cli/acs.
+* Remove oidc issuer related code after GA, reuse the implementation in azure-cli/acs.
+* Fix monitoring addon option `--enable-syslog` for `aks addon enable`.
+* Remove deprecated option `--node-zones`, use `--zones` instead.
+* Remove gpu instance profile related code after GA, reuse the implementation in azure-cli/acs.
+* Remove http proxy config related code after GA, reuse the implementation in azure-cli/acs.
 
 0.5.120
 +++++++
+
 * Remove file, blob csi driver and snapshot controller related CSI driver code after GA, reuse the implementation in azure-cli/acs.
 * Remove Azure Dedicated Host related code after GA, reuse the implementation in azure-cli/acs.
 * Remove KMS related code after GA, reuse the implementation in azure-cli/acs.